+++ packages/extract-svg-sprite-webpack-plugin/lib/plugin.js
@@ -10,11 +10,11 @@ const Replacer = require('./utils/replacer');
 let INSTANCE_COUNTER = 0;
 
 class ExtractSvgSpritePlugin {
-  static extract(options) {
+  static loader(options) {
     return { loader: LOADER_PATH, options };
   }
 
-  static extractFromCss() {
+  static cssLoader() {
     return { loader: CSS_LOADER_PATH };
   }
 
@@ -28,37 +28,15 @@ class ExtractSvgSpritePlugin {
     return NAMESPACE;
   }
 
-  addSymbol(symbol) {
-    this.compiler.addSymbol(symbol);
-  }
-
-  isExtractTextPluginCompiler(compiler) {
-    return (
-      compiler.name && compiler.name.startsWith('extract-text-webpack-plugin')
-    );
-  }
-
-  isMiniCssExtractPlugin(compiler) {
-    return (
-      compiler.name && compiler.name.startsWith('mini-css-extract-plugin')
-    );
-  }
-
-  isHtmlPluginCompiler(compiler) {
-    return (
-      compiler.name && compiler.name.startsWith('html-webpack-plugin')
-    );
-  }
-
   apply(compiler) {
     // TODO refactor this ugly way to avoid double compilation when using extract-text-webpack-plugin
     let prevResult;
     // eslint-disable-next-line arrow-body-style
-    const compileSprites = () => {
+    const compileSprites = compilation => {
       return (
         prevResult
           ? Promise.resolve(prevResult)
-          : this.compiler.compile()
+          : this.compiler.compile(compilation)
       ).then(result => {
         prevResult = result;
         return result;
@@ -71,20 +49,23 @@ class ExtractSvgSpritePlugin {
           .tap(NAMESPACE, loaderCtx => this.hookNormalModuleLoader(loaderCtx));
 
         compilation.hooks.additionalAssets
-          .tapPromise(NAMESPACE, () => compileSprites()
+          .tapPromise(NAMESPACE, () => compileSprites(compilation)
             .then(result => this.hookAdditionalAssets(compilation, result)));
       });
 
       compiler.hooks.compilation.tap(NAMESPACE, compilation => {
-        if (this.isMiniCssExtractPlugin(compilation.compiler)) {
+        if (
+          compilation.compiler.name &&
+          compilation.compiler.name.startsWith('mini-css-extract-plugin')
+        ) {
           compilation.hooks.additionalAssets
-            .tapPromise(NAMESPACE, () => compileSprites()
+            .tapPromise(NAMESPACE, () => compileSprites(compilation)
               .then(result => this.hookAdditionalAssets(compilation, result)));
         }
 
         if (compilation.hooks.htmlWebpackPluginBeforeHtmlGeneration) {
           compilation.hooks.htmlWebpackPluginBeforeHtmlGeneration
-            .tapAsync(NAMESPACE, (htmlPluginData, done) => compileSprites()
+            .tapAsync(NAMESPACE, (htmlPluginData, done) => compileSprites(compilation)
               .then(result => {
                 this.hookBeforeHtmlGeneration(htmlPluginData, result);
                 done(null, htmlPluginData);
@@ -93,7 +74,10 @@ class ExtractSvgSpritePlugin {
       });
     } else {
       compiler.plugin('compilation', compilation => {
-        if (this.isHtmlPluginCompiler(compilation.compiler)) {
+        if (
+          compilation.compiler.name &&
+          compilation.compiler.name.startsWith('html-webpack-plugin')
+        ) {
           return;
         }
 
@@ -102,14 +86,14 @@ class ExtractSvgSpritePlugin {
           loaderCtx => this.hookNormalModuleLoader(loaderCtx)
         );
 
-        compilation.plugin('additional-assets', done => compileSprites().then(result => {
+        compilation.plugin('additional-assets', done => compileSprites(compilation).then(result => {
           this.hookAdditionalAssets(compilation, result);
           done();
         }));
 
         compilation.plugin(
           'html-webpack-plugin-before-html-generation',
-          (htmlPluginData, done) => compileSprites().then(result => {
+          (htmlPluginData, done) => compileSprites(compilation).then(result => {
             this.hookBeforeHtmlGeneration(htmlPluginData, result);
             done(null, htmlPluginData);
           })
@@ -131,7 +115,7 @@ class ExtractSvgSpritePlugin {
       });
 
       if (filename) {
-        compilation.assets[filename] = {
+        compilation.assets[filename.split('?')[0]] = {
           source: () => content,
           size: () => content.length
         };