+++ packages/extract-svg-sprite-webpack-plugin/lib/utils/replacement-generator.js
@@ -3,10 +3,8 @@ const { getHashDigest } = require('loader-utils');
 
 const { NAMESPACE } = require('../config');
 
-const TOKEN_START = '___';
-const TOKEN_END = '___';
-
-const REPLACEMENTS = {
+const Replacements = {
+  PUBLIC_PATH: 'public_path',
   SPRITE_FILENAME: 'sprite_filename',
   SYMBOL_BG_POSITION_LEFT: 'symbol_bg_position_left',
   SYMBOL_BG_POSITION_TOP: 'symbol_bg_position_top',
@@ -14,80 +12,103 @@ const REPLACEMENTS = {
   SYMBOL_BG_SIZE_HEIGHT: 'symbol_bg_size_height'
 };
 
-function generate(id, replacementName = '') {
+/**
+ * @param {string} id
+ * @return {string}
+ */
+function generateToken(id) {
   const value = [
     NAMESPACE,
-    id,
-    replacementName && `:${replacementName}`
+    id
   ].filter(Boolean).join('');
 
-  return `${TOKEN_START}${getHashDigest(value)}${TOKEN_END}`;
+  return `___${getHashDigest(value)}___`;
 }
 
-module.exports = class Generator {
-  static symbolRequest(symbol, config) {
-    const { filename, emit } = config;
+class Replacement {
+  constructor(token, replaceTo) {
+    this.token = generateToken(token);
+    this.replaceTo = replaceTo;
+  }
+}
+
+class ReplacementGenerator {
+  /**
+   * @param {SpriteSymbol} symbol
+   * @param {Object} config
+   * @param {string} config.filename
+   * @param {boolean} config.emit
+   * @param {string} config.spriteType
+   * @return {Replacement}
+   */
+  static symbolUrl(symbol, config) {
+    const { filename, emit, spriteType } = config;
     let replaceTo;
 
     if (!filename || !emit) {
       replaceTo = `#${symbol.id}`;
     } else {
-      replaceTo = config.spriteType === mixer.StackSprite.TYPE
+      replaceTo = spriteType === mixer.StackSprite.TYPE
         ? `${filename}#${symbol.id}`
         : filename;
     }
 
-    return {
-      value: generate(REPLACEMENTS.SPRITE_FILENAME, symbol.request),
+    return new Replacement(
+      `${Replacements.SPRITE_FILENAME}:${symbol.request}`,
       replaceTo
-    };
+    );
   }
 
-  static bgPosLeft(request, position) {
-    return {
-      value: generate(REPLACEMENTS.SYMBOL_BG_POSITION_LEFT, request),
-      replaceTo: position ? position.bgPosition.left : undefined
-    };
+  /**
+   * @param {string} symbolUrl
+   * @param {SpriteSymbolPosition} position
+   * @return {Replacement}
+   */
+  static bgPosLeft(symbolUrl, position) {
+    return new Replacement(
+      `${Replacements.SYMBOL_BG_POSITION_LEFT}:${symbolUrl}`,
+      position ? position.bgPosition.left : undefined
+    );
   }
 
-  static bgPosTop(request, position) {
-    return {
-      value: generate(REPLACEMENTS.SYMBOL_BG_POSITION_TOP, request),
-      replaceTo: position ? position.bgPosition.top : undefined
-    };
+  /**
+   * @param {string} symbolUrl
+   * @param {SpriteSymbolPosition} position
+   * @return {Replacement}
+   */
+  static bgPosTop(symbolUrl, position) {
+    return new Replacement(
+      `${Replacements.SYMBOL_BG_POSITION_TOP}:${symbolUrl}`,
+      position ? position.bgPosition.top : undefined
+    );
   }
 
-  static bgSizeWidth(request, position) {
-    return {
-      value: generate(REPLACEMENTS.SYMBOL_BG_SIZE_WIDTH, request),
-      replaceTo: position ? position.bgSize.width : undefined
-    };
+  /**
+   * @param {string} symbolUrl
+   * @param {SpriteSymbolPosition} position
+   * @return {Replacement}
+   */
+  static bgSizeWidth(symbolUrl, position) {
+    return new Replacement(
+      `${Replacements.SYMBOL_BG_SIZE_WIDTH}:${symbolUrl}`,
+      position ? position.bgSize.width : undefined
+    );
   }
 
-  static bgSizeHeight(request, position) {
-    return {
-      value: generate(REPLACEMENTS.SYMBOL_BG_SIZE_HEIGHT, request),
-      replaceTo: position ? position.bgSize.height : undefined
-    };
+  /**
+   * @param {string} symbolUrl
+   * @param {SpriteSymbolPosition} position
+   * @return {Replacement}
+   */
+  static bgSizeHeight(symbolUrl, position) {
+    return new Replacement(
+      `${Replacements.SYMBOL_BG_SIZE_HEIGHT}:${symbolUrl}`,
+      position ? position.bgSize.height : undefined
+    );
   }
+}
 
-  static symbol({ symbol, position, config }) {
-    const { request } = symbol;
-
-    // css replacements
-    const replacements = [
-      Generator.symbolRequest(symbol, config),
-      Generator.bgPosLeft(request, position),
-      Generator.bgPosTop(request, position),
-      Generator.bgSizeWidth(request, position),
-      Generator.bgSizeHeight(request, position)
-    ];
-
-    return replacements.reduce((acc, replacement) => {
-      acc[replacement.value] = replacement.replaceTo;
-      return acc;
-    }, {});
-  }
-};
-
-module.exports.REPLACEMENTS = REPLACEMENTS;
+module.exports = ReplacementGenerator;
+module.exports.Replacement = Replacement;
+module.exports.generateToken = generateToken;
+module.exports.Replacements = Replacements;