+++ packages/posthtml-transform/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.2.0"></a>
+# [0.2.0](https://github.com/kisenka/svg-mixer/packages/posthtml-transform/compare/posthtml-transform@0.1.2...posthtml-transform@0.2.0) (2018-07-23)
+
+
+### Features
+
+* add possibility to convert alpha colors to rgb ([dea0807](https://github.com/kisenka/svg-mixer/packages/posthtml-transform/commit/dea0807))
+
+
+
+
 <a name="0.1.2"></a>
 ## [0.1.2](https://github.com/kisenka/svg-mixer/packages/posthtml-transform/compare/posthtml-transform@0.1.1...posthtml-transform@0.1.2) (2018-05-11)