+++ packages/svg-mixer/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.3.2"></a>
+## [2.3.2](https://github.com/kisenka/svg-mixer/packages/svg-mixer/compare/svg-mixer@2.3.1...svg-mixer@2.3.2) (2018-07-20)
+
+
+
+
+**Note:** Version bump only for package svg-mixer
+
 <a name="2.3.1"></a>
 ## [2.3.1](https://github.com/kisenka/svg-mixer/packages/svg-mixer/compare/svg-mixer@2.3.0...svg-mixer@2.3.1) (2018-06-18)